change permission in model
